IndustryCalculationServiceImpl progress
It's doing most of the work of what I want so far.  With the invention and manufacturing calculators out of the way now I need to work on the overhead service implementation and definition.  These will actually probably end up being the hardest but for the short term I'll change the FreightDetails model object to contain a Map of prices and calculate to and from Amarr -> Atreen, Jita -> Atreen, Dodixie -> Atreen and Rens -> Atreen.  V1 will only end up being useful of Lucifer's Hammer but minor refactorings later should be fairly straight forward to make it more general purpose.