Working towards a more performant backend
Load all of the target regions orders and cache it, sort through it using streams to find the data that we want.

Unit tests to back this behavior up (hopefully its actually more performant).  An individual call might not be but multiples or hundreds over a short period of time (ttl on the cache) should indeed be much faster as the entirety of the market data is in cache.