Failed experiment
The getAllMarketOrders method in the CrestMarketOrderFetcher would not cache (I was hoping spring fixed this) due to the same class calling its own method (and not going through the spring proxy).  Maybe I'm doing something wrong and there is a way for the proxy to intercept this method call on itself.  First call would be absolutely horrible but with a long enough cache life we could reap some awesome benefits.  I'll consider making a sub component that this service can depend on for this caching functionality, at least for backend analysis.

Ideally I'd like to get to a place where the backend is checking every single possible item as often as possible (or at least a few times per day) and this string of checks would benefit greatly from such a convenient caching of ALL market orders in a or a few regions.