Load Balancing router for industry-calculator-workers
This is basically stolen verbatim from the lbbroker.java jeromq example.  I should dive deeper to fully understand what is happening as there will be more happening here.  Such as connecting to the industry-director for directions on what should be dove into when diving down build chains and anything else that this thing might need.