Router, worker, client initial implementation
Still need to write unit tests.  These have been tested to work with the LocalRouterCommunicationApp in test/java/com/aba/industry/router/client/impl package of the industry-calculator-router-client module.  Setting anything less than what is in that app will cause the worker to crash with an NPE currently.  Next steps, unit tests, harden the worker so it doesn't lose it's mind so easily, refactor the BuildCalculationResult or return a wrapped object that has a status field and some kind of other debug information.

Tomorrow I can start to maybe dive into the meat of the Slack Bot and get some rudimentary stuff implemented so we can start to beat it up and see how badly it fails :D.