Rethink... you're not mature enough
It's not me... it's you.  While I thought we could get along and be good bud's it was just not meant to be :(.  Your integration with the largest enterprise (and now hobby) framework was just not good enough.  I want a programming model I'm used to... that makes sense... that makes me NOT WRITE A CRAP TON OF BOILER PLATE GARBAGE!!!  I'll check you out some other year :(.