#15 Various Workings
I really need to commit more often.  Looks like I was working on the warehouse, have the beginnings of something to check the auto build checker output and have started to fine tune the build checker a bit.  Can't remember what else might be in this commit.

Will definitely need to test and play with the auto build checker a bit more... IntelliJ is throwing some errors about a ConnectionFactory bean not existing... maybe I was tweaking that?